refactor: add output and latexdiff managers
